:recycle: api: move sessions import route
