+++ api/lib/application/certification-centers/certification-center-controller.js
@@ -11,6 +11,7 @@ const sessionSerializer = require('../../infrastructure/serializers/jsonapi/sess
 
 const queryParamsUtils = require('../../infrastructure/utils/query-params-utils');
 const map = require('lodash/map');
+const csvHelpers = require('../../../scripts/helpers/csvHelpers');
 
 module.exports = {
   async saveSession(request) {
@@ -167,5 +168,11 @@ module.exports = {
       return h.response(serializedCertificationCenterInvitation).created();
     }
     return h.response(serializedCertificationCenterInvitation);
+  },
+
+  async importSessions(request, h) {
+    const data = await csvHelpers.parseCsvWithHeader(request.payload.file.path);
+    await usecases.createSessions({ data });
+    return h.response().code(200);
   },
 };
