+++ api/lib/application/certification-centers/index.js
@@ -170,6 +170,25 @@ exports.register = async function (server) {
         tags: ['api', 'certification-center'],
       },
     },
+    {
+      method: 'POST',
+      path: '/api/certification-centers/{certificationCenterId}/sessions/import',
+      config: {
+        handler: certificationCenterController.importSessions,
+        payload: {
+          maxBytes: 20715200,
+          output: 'file',
+          parse: true,
+          allow: 'multipart/form-data',
+          multipart: true,
+        },
+        tags: ['api', 'certification-center', 'sessions'],
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+            "- Elle permet d'importer un fichier contenant une liste de sessions à créer",
+        ],
+      },
+    },
   ];
 
   server.route([
