+++ api/lib/application/sessions/index.js
@@ -798,25 +798,6 @@ exports.register = async (server) => {
         ],
       },
     },
-    {
-      method: 'POST',
-      path: '/api/sessions/import',
-      config: {
-        handler: sessionController.importSessions,
-        payload: {
-          maxBytes: 20715200,
-          output: 'file',
-          parse: true,
-          allow: 'multipart/form-data',
-          multipart: true,
-        },
-        tags: ['api', 'sessions'],
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            "- Elle permet d'importer un fichier contenant une liste de sessions à créer",
-        ],
-      },
-    },
   ]);
 };
