+++ api/tests/acceptance/application/certification-centers/certification-centers-controller-post-import-sessions_test.js
@@ -5,23 +5,24 @@ const fs = require('node:fs');
 const { stat } = require('node:fs/promises');
 const streamToPromise = require('stream-to-promise');
 
-describe('Acceptance | Controller | session-controller-post-import-sessions', function () {
+describe('Acceptance | Controller | certification-centers-controller-post-import-sessions', function () {
   let server;
 
   beforeEach(async function () {
     server = await createServer();
   });
 
-  describe('POST /api/sessions/import', function () {
+  describe('POST /api/certification-centers/{certificationCenterId}/sessions/import', function () {
     context('when user imports sessions', function () {
       it('should return status 200', async function () {
         // given
         const userId = databaseBuilder.factory.buildUser().id;
+        const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
         await databaseBuilder.commit();
 
-        const csvFileName = 'files/sessions-import/sessions.csv';
+        const csvFileName = 'files/sessions.csv';
         const csvFilePath = `${__dirname}/${csvFileName}`;
-        const options = await createRequest({ csvFilePath, userId });
+        const options = await createRequest({ csvFilePath, userId, certificationCenterId });
 
         // when
         const response = await server.inject(options);
@@ -33,7 +34,7 @@ describe('Acceptance | Controller | session-controller-post-import-sessions', fu
   });
 });
 
-async function createRequest({ csvFilePath, userId }) {
+async function createRequest({ csvFilePath, userId, certificationCenterId }) {
   const form = new FormData();
   const knownLength = await stat(csvFilePath).size;
   form.append('file', fs.createReadStream(csvFilePath), { knownLength });
@@ -43,7 +44,7 @@ async function createRequest({ csvFilePath, userId }) {
   const headers = Object.assign({}, form.getHeaders(), { authorization: `Bearer ${token}` });
   return {
     method: 'POST',
-    url: `/api/sessions/import`,
+    url: `/api/certification-centers/${certificationCenterId}/sessions/import`,
     headers,
     payload,
   };
