+++ certif/tests/unit/components/sessions/panel-header_test.js
@@ -2,6 +2,7 @@ import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
 import createGlimmerComponent from '../../../helpers/create-glimmer-component';
 import sinon from 'sinon';
+import Service from '@ember/service';
 
 module('Unit | Component | panel-header', function (hooks) {
   setupTest(hooks);
@@ -66,6 +67,16 @@ module('Unit | Component | panel-header', function (hooks) {
   module('#importSessions', function () {
     test('should call upload with the right parameters', async function (assert) {
       // given
+      const store = this.owner.lookup('service:store');
+      const currentAllowedCertificationCenterAccess = store.createRecord('allowed-certification-center-access', {
+        id: 123,
+      });
+
+      class CurrentUserStub extends Service {
+        currentAllowedCertificationCenterAccess = currentAllowedCertificationCenterAccess;
+      }
+
+      this.owner.register('service:current-user', CurrentUserStub);
       const token = 'a token';
 
       component.session = {
@@ -87,7 +98,7 @@ module('Unit | Component | panel-header', function (hooks) {
       // then
       sinon.assert.calledOnce(component.notifications.success);
       assert.ok(
-        file.upload.calledWith('/api/sessions/import', {
+        file.upload.calledWith('/api/certification-centers/123/sessions/import', {
           headers: { Authorization: `Bearer ${token}` },
         })
       );
@@ -96,6 +107,17 @@ module('Unit | Component | panel-header', function (hooks) {
     test('should call the notifications service in case of an error', async function (assert) {
       // given
       const token = 'a token';
+      const store = this.owner.lookup('service:store');
+      const currentAllowedCertificationCenterAccess = store.createRecord('allowed-certification-center-access', {
+        id: 123,
+      });
+
+      class CurrentUserStub extends Service {
+        currentAllowedCertificationCenterAccess = currentAllowedCertificationCenterAccess;
+      }
+
+      this.owner.register('service:current-user', CurrentUserStub);
+
       component.session = {
         isAuthenticated: true,
         data: {
